+++ mcu/attiny-hal/src/attiny2313.rs
@@ -23,3 +23,9 @@ pub mod port {
         }
     }
 }
+
+pub mod simple_pwm {
+    pub use crate::periphals::simple_pwm::*;
+
+    // Fixme: Implement PWM for ATtiny2313.
+}
