+++ mcu/attiny-hal/src/periphals/mod.rs
@@ -3,3 +3,4 @@
 pub mod adc;
 pub mod eeprom;
 pub mod port;
+pub mod simple_pwm;
