+++ mcu/attiny-hal/src/periphals/simple_pwm.rs
@@ -1,0 +1,1 @@
+pub use avr_hal_generic::simple_pwm::{IntoPwmPin, Prescaler, PwmPinOps};
